[Dependencies] WIP resolving dependencies
Can read the dependencies from the Cult.toml.

This involved some re-purposing of converting the semantiv version of the
package into one that is usable for the dependencies.

For now storing dependencies in a Map such that the identifier of the dependency
maps to the version. So the dependency declared like

org.apache.ivy.ivy = "2.5.2"

would map to

{org.apache.ivy.ivy=Version[major=2, minor=5, patch=2]}

Going to need a better way to distinguish between the group and artifact id, but
this is a WIP.