[Dependencies] Move away from Ivy
Ivy didn't work when building a native image, so went with a simpler implementation that
tries to fetch a dependency from a Maven repository. Doesn't handle any transitive dependency
fetching, but maybe would add some of that later or figure out how to make something Ivy-like
work with the native-image stuff.