[Build] Can build using a dependency with a local path
Can declare dependencies like `org.cult_cult = { path = "../../cult" }` that are
paths to other local Cult packages. If the package was built already, then
will use the package's library jar when building.

The classes build ok, but need to work on the testing code next to fix up the
reflection and classloading so that they run the project being builts tests (and
not the ones bundled with the Cult jar :P)